backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.1
sqlalchemy==2.0.30
aiosqlite==0.20.0
greenlet==3.0.3
httpx==0.27.0
google-generativeai==0.7.2
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
websockets==12.0
Pillow==10.3.0
